# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prediction League 26/27",
    "description": "Premier League predictions, FPL and a watch log for four friends. Every response is served from cache; the poller is the only process that talks to an upstream.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/session": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Sign In",
        "description": "One field, one code word, compared constant-time.",
        "operationId": "sign_in_api_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "session"
        ],
        "summary": "Sign Out",
        "operationId": "sign_out_api_session_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "tags": [
          "session"
        ],
        "summary": "Me",
        "operationId": "me_api_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginIn": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "LoginIn"
      },
      "MeOut": {
        "properties": {
          "person": {
            "$ref": "#/components/schemas/PersonOut"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonOut"
            },
            "type": "array",
            "title": "People"
          },
          "season": {
            "type": "string",
            "title": "Season"
          },
          "prediction_lock": {
            "type": "string",
            "format": "date-time",
            "title": "Prediction Lock"
          },
          "locked": {
            "type": "boolean",
            "title": "Locked"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          }
        },
        "type": "object",
        "required": [
          "person",
          "people",
          "season",
          "prediction_lock",
          "locked",
          "server_time"
        ],
        "title": "MeOut"
      },
...
      "PersonOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "fpl_entry_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fpl Entry Id"
          }
        },
        "type": "object",
        "required": [
          "key",
          "name",
          "city",
          "timezone",
          "country",
          "fpl_entry_id"
        ],
        "title": "PersonOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}